Fixed cast related issue with query.
git-svn-id: http://svn.code.sf.net/p/postbooks/code/xtuple/trunk@6701 ea1a15ef-c238-0410-93cb-fa28e9f07733

+++ guiclient/returnAuthorization.cpp
@@ -845,7 +845,7 @@ void returnAuthorization::sCheckAuthorizationNumber()
     query.prepare( "SELECT rahead_id "
                    "FROM rahead "
                    "WHERE (rahead_number=:rahead_number)" );
-    query.bindValue(":rahead_number", _authNumber->text().toInt());
+    query.bindValue(":rahead_number", _authNumber->text());
     query.exec();
     if (query.first())
     {
